DataGenerator

+++ src/main.rs
@@ -2,8 +2,8 @@ mod data_generator;
 mod treap;
 mod dynamic_array;
 
-type ID = u128;
-type Key = u128;
+type ID = u64;
+type Key = u32;
 type Element = (ID, Key);
 
 fn main() {
